tools/fanqie_scraper.py
import logging
from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)

class FanqieScraper:
    def __init__(self, config):
        self.config = config

    def scrape_top_rankings(self):
        logger.info("Scraping Fanqie novel rankings...")
        # Placeholder for scraping logic (requires playwright or requests depending on dynamic rendering)
        # Returns a list of novel IDs and their ranks
        return [{"novel_id": "123456", "rank": 1, "title": "Example Top Novel"}]

    def read_entire_novel(self, novel_id):
        logger.info(f"Reading entire novel {novel_id} for deep learning...")
        # Placeholder: paginate through chapters, extract text
        return f"Full text data for {novel_id} demonstrating character growth and plot construction."

    def read_comments(self, novel_id):
        logger.info(f"Reading audience comments for novel {novel_id}")
        return ["Great character development!", "The plot in chapter 5 was slow."]
