requirements.txt
openclaw>=0.1.0
litellm>=1.0.0
pyyaml>=6.0
playwright>=1.0.0
playwright-stealth>=1.0.0
